Add files via upload
The objective of this project is to enhance customer experience and sell  by knowing the period of sales. This project will consider data analysis and visualization in Python with various libraries such as Pandas, Seaborn, and Matplotlib to figure out the peak sale period. The key takeaways can then be used directly in controlling the inventory and marketing strategy in a way that products are available and promotions can be efficient when they are in demand. Eventually, operation can be optimized and revenue growth can be generated with an increase in customer satisfaction.